acl incompleto

+++ app/Http/Controllers/AnuncioController.php
@@ -9,7 +9,10 @@ use App\Http\Requests\SyncAnuncioTotemRequest;
 use App\Http\Requests\UpdateAnuncioRequest;
 use App\Models\Anuncio;
 use App\Services\AnuncioService;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
+use Tymon\JWTAuth\Facades\JWTAuth;
 
 class AnuncioController extends Controller
 {
@@ -30,9 +33,16 @@ class AnuncioController extends Controller
     public function index()
     {
         try {
+            $this->authorize('viewAny', $this->anuncio);
             $anuncios = $this->anuncio->paginate();
             return response()->json($anuncios);
-        } catch (\Exception) {
+        } catch (\Exception $e) {
+            if ($e instanceof AuthorizationException) {
+                return response()->json([
+                    'exception' => 'Unauthorized',
+                    'message' => $e->getMessage(),
+                ], 401);
+            }
             throw new InternalServerErrorException();
         }
     }
@@ -46,9 +56,16 @@ class AnuncioController extends Controller
     public function store(CreateAnuncioRequest $request)
     {
         try {
+            $this->authorize('create', User::class);
             $anuncio = $this->anuncioService->create($request->except('totem_id'), $request['totem_id']);
             return response()->json($anuncio, 201);
-        } catch (\Exception) {
+        } catch (\Exception $e) {
+            if ($e instanceof AuthorizationException) {
+                return response()->json([
+                    'exception' => 'Unauthorized',
+                    'message' => $e->getMessage(),
+                ], 401);
+            }
             throw new InternalServerErrorException();
         }
     }
@@ -62,6 +79,7 @@ class AnuncioController extends Controller
     public function vincularAnuncioTotem(SyncAnuncioTotemRequest $request)
     {
         try {
+            $this->authorize('sync', $this->anuncio);
             $syncSuccess = $this->anuncioService->syncAnuncioTotem($request->validated());
             if ($syncSuccess === 404) {
                 return response()->json([
@@ -75,6 +93,12 @@ class AnuncioController extends Controller
                 ], 201);
             }
         } catch (\Exception $e) {
+            if ($e instanceof AuthorizationException) {
+                return response()->json([
+                    'exception' => 'Unauthorized',
+                    'message' => $e->getMessage(),
+                ], 401);
+            }
             throw new InternalServerErrorException();
         }
     }
@@ -88,6 +112,7 @@ class AnuncioController extends Controller
     public function show($id)
     {
         try {
+            $this->authorize('view', $this->anuncio->find($id));
             $anuncio = $this->anuncio->find($id);
             if ($anuncio == null) {
                 return response()->json([
@@ -95,9 +120,14 @@ class AnuncioController extends Controller
                     'message' => 'Não foi encontrado nenhum anúncio com este ID',
                 ], 404);
             }
-
             return response()->json($anuncio);
-        } catch (\Exception) {
+        } catch (\Exception $e) {
+            if ($e instanceof AuthorizationException) {
+                return response()->json([
+                    'exception' => 'Unauthorized',
+                    'message' => $e->getMessage(),
+                ], 401);
+            }
             throw new InternalServerErrorException();
         }
     }
@@ -112,6 +142,7 @@ class AnuncioController extends Controller
     public function update(UpdateAnuncioRequest $request, $id)
     {
         try {
+            $this->authorize('update', $this->anuncio->find($id));
             $anuncio = $this->anuncioService->update($request, $id);
             if ($anuncio == '404') {
                 return response()->json([
@@ -120,7 +151,13 @@ class AnuncioController extends Controller
                 ], 404);
             }
             return response()->json($anuncio);
-        } catch (\Exception) {
+        } catch (\Exception $e) {
+            if ($e instanceof AuthorizationException) {
+                return response()->json([
+                    'exception' => 'Unauthorized',
+                    'message' => $e->getMessage(),
+                ], 401);
+            }
             throw new InternalServerErrorException();
         }
     }
@@ -134,18 +171,25 @@ class AnuncioController extends Controller
     public function destroy($id)
     {
         try {
+            $this->authorize('delete', $this->anuncio);
             $anuncio = $this->anuncio->find($id);
             if ($anuncio == null) {
                 return response()->json([
                     'except' => 'NotFoundException',
-                    'message' => 'Não foi encontrado nenhum anúncio',
+                    'message' => 'Não foi encontrado nenhum anúncio com este ID.',
                 ], 404);
             }
             $anuncio->delete();
             return response()->json([
                 'message' => 'Anúncio Excluído com sucesso'
             ], 200);
-        } catch (\Exception) {
+        } catch (\Exception $e) {
+            if ($e instanceof AuthorizationException) {
+                return response()->json([
+                    'exception' => 'Unauthorized',
+                    'message' => $e->getMessage(),
+                ], 401);
+            }
             throw new InternalServerErrorException();
         }
     }
@@ -186,6 +230,30 @@ class AnuncioController extends Controller
             $anuncios = Anuncio::where('nome_anunciante', 'like', '%' .  $nomeAnunciante . '%')->paginate();
             return response()->json($anuncios);
         } catch (\Exception $e) {
+            throw new InternalServerErrorException();
+        }
+    }
+
+    public function anunciosAtivos()
+    {
+        try {
+            $anuncios = $this->anuncio->where('ativo', 1)->count('*');
+            return response()->json([
+                'anuncios' => $anuncios
+            ]);
+        } catch (\Exception $e) {
+            throw new InternalServerErrorException();
+        }
+    }
+
+    public function faturamento()
+    {
+        try {
+            $faturamento = $this->anuncioService->faturamento();
+            return response()->json([
+                'faturamento' => $faturamento
+            ]);
+        } catch (\Exception) {
             throw new InternalServerErrorException();
         }
     }
